Adding small hybrid design
This replicates the size 14 Cannondale CAAD3 from 1996, although the
"effective" top tube length quoted there is too long for the quoted
wheelbase and front centre. This design matches the wheelbase and front
centre. The top tube angle was measured from one of the pictures but I
don't know what size that was.